Refactor UIState, add NetworkState
