+++ app/src/main/java/pl/damrad/marvelcomicsapp/fragments/ComicsFragment.kt
@@ -2,7 +2,6 @@ package pl.damrad.marvelcomicsapp.fragments
 
 import android.graphics.Color
 import android.os.Bundle
-import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
@@ -18,7 +17,8 @@ import pl.damrad.marvelcomicsapp.adapters.PaginationScrollListener
 import pl.damrad.marvelcomicsapp.databinding.FragmentComicsBinding
 import pl.damrad.marvelcomicsapp.other.ComicItemCreator
 import pl.damrad.marvelcomicsapp.other.Key
-import pl.damrad.marvelcomicsapp.other.UIState
+import pl.damrad.marvelcomicsapp.states.NetworkState
+import pl.damrad.marvelcomicsapp.states.UIState
 import pl.damrad.marvelcomicsapp.viewmodels.MainViewModel
 
 class ComicsFragment : Fragment() {
@@ -58,16 +58,16 @@ class ComicsFragment : Fragment() {
     private fun setObservers() {
         mainViewModel.connectionState.observe(viewLifecycleOwner) { state ->
             when (state) {
-                is UIState.Connected -> {
+                is NetworkState.Connected -> {
                     if (connectionBar.isShown) connectionBar.dismiss()
                 }
-                is UIState.Disconnected -> {
+                is NetworkState.Disconnected -> {
                     if (!connectionBar.isShown) connectionBar.show()
                 }
-                is UIState.Warning -> {
+                is NetworkState.Warning -> {
                     Toasty.warning(requireContext(), getString(R.string.something_went_wrong), Toasty.LENGTH_LONG).show()
                 }
-                is UIState.Timeout -> {
+                is NetworkState.Timeout -> {
                     Toasty.warning(requireContext(), getString(R.string.connection_time_out), Toasty.LENGTH_LONG).show()
                 }
             }
