--- app/src/main/java/pl/damrad/marvelcomicsapp/other/UIState.kt
@@ -1,17 +1,0 @@
-package pl.damrad.marvelcomicsapp.other
-
-import android.os.Bundle
-
-sealed class UIState {
-    object Initialized : UIState()
-    object Refreshing : UIState()
-    object InProgress : UIState()
-    object Connected: UIState()
-    object Disconnected: UIState()
-    object Timeout: UIState()
-    object Success : UIState()
-    object Error : UIState()
-    object Warning : UIState()
-    data class ErrorResponse(val text: String? = null) : UIState()
-    data class NavigateTo(val key: String?, val bundle: Bundle? = null) : UIState()
-}
