+++ app/src/main/java/pl/damrad/marvelcomicsapp/states/NetworkState.kt
@@ -1,0 +1,8 @@
+package pl.damrad.marvelcomicsapp.states
+
+sealed class NetworkState {
+    object Connected: NetworkState()
+    object Disconnected: NetworkState()
+    object Timeout: NetworkState()
+    object Warning : NetworkState()
+}
